updated storage + jsconfig
