+++ jsconfig.json
@@ -1,11 +1,20 @@
 {
+  "include": [
+    "./src/**/*"
+  ],
+  "module": "commonjs",
   "compilerOptions": {
+    "target": "es6",
     "baseUrl": ".",
     "paths": {
-      "@": ["./src"]
+      "@/*": [
+        "./src/*"
+      ]
     }
   },
   "typeAcquisition": {
-    "include": ["jest"]
+    "include": [
+      "jest"
+    ]
   }
 }
