+++ src/lib/addSite.test.js
@@ -1,10 +1,10 @@
 import addSite from './addSite';
-import storage from './storage';
+import storage from '@/lib/storage';
 import { STORE_WEBSITES, STORE_WEBSITES_INDEX } from '@/constants/storeKey';
 import * as swr from 'swr';
 import makeId from './makeId';
 
-jest.mock('./storage');
+jest.mock('@/lib/storage');
 jest.mock('./makeId');
 jest.mock('swr');
