+++ src/lib/addTask.ts
@@ -1,5 +1,5 @@
 import { mutate } from 'swr';
-import storage from './storage';
+import storage from '@/lib/storage';
 import makeId from './makeId';
 import makeLogger from './makeLogger';
 import { STORE_TASKS, STORE_TASKS_INDEX } from '@/constants/storeKey';
